+++ src/vec2.cpp
@@ -1,6 +1,6 @@
-#include "../include/vec2.h"
+#include "vec2.h"
 
-double vec2::dot(const vec2& oth) const 
+float vec2::dot(const vec2& oth) const
 {
     return x * oth.x + y * oth.y;
 }
